Revert "This is The Recording Sound"
This reverts commit 4d1ca7dcb457357dbc25481b64e101d8042a1e77.
