+++ MySaber/app/src/main/java/hillnerds/me/mysaber/MainActivity.java
@@ -1,181 +1,77 @@
 package hillnerds.me.mysaber;
 
-import android.Manifest;
-import android.app.Dialog;
-import android.content.Context;
-import android.content.pm.PackageManager;
-import android.media.MediaPlayer;
-import android.media.MediaRecorder;
+import android.media.Image;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.ActivityCompat;
-import android.support.v7.app.AlertDialog;
+import android.support.design.widget.FloatingActionButton;
+import android.support.design.widget.Snackbar;
 import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.View;
-import android.view.ViewGroup;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.widget.Button;
-import android.widget.EditText;
-import android.widget.LinearLayout;
-import android.widget.Toast;
+import android.widget.ImageButton;
+import android.widget.ImageView;
 
-import java.io.IOException;
-
-import static hillnerds.me.mysaber.R.id.btnRec1;
+import static android.view.View.*;
 
 public class MainActivity extends AppCompatActivity {
 
-    private static final String LOG_TAG = "AudioRecordTest";
-    private static final int REQUEST_RECORD_AUDIO_PERMISSION = 200;
-    private static String mFileName = null;
-
-    private RecordButton mRecordButton = null;
-
-    private MediaRecorder mRecorder = null;
-    private boolean recordable = true;
-
-    // Requesting permission to RECORD_AUDIO
-    private boolean permissionToRecordAccepted = false;
-    private String [] permissions = {Manifest.permission.RECORD_AUDIO};
-
     @Override
-    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        switch (requestCode){
-            case REQUEST_RECORD_AUDIO_PERMISSION:
-                permissionToRecordAccepted  = grantResults[0] == PackageManager.PERMISSION_GRANTED;
-                break;
-        }
-        if (!permissionToRecordAccepted ) finish();
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
 
+        final ImageView imBlade1 = (ImageView) findViewById(R.id.imageView);
+        final ImageView imBlade2 = (ImageView) findViewById(R.id.imageView);
 
-}
-    private void onRecord(boolean start) {
-        if (start) {
-            mFileName = getExternalCacheDir().getAbsolutePath();
-            mFileName += "FunFile";
-            mFileName += ".3gp";
+        Button btnRec1 = (Button) findViewById(R.id.btnRec1);
 
-            startRecording();
-        } else {
-            stopRecording();
-        }
-    }
-
-    public String userFile(){
-
-        final String[] accepted = new String[1];
-        final AlertDialog.Builder mBuilder = new AlertDialog.Builder(MainActivity.this);
-        View tveiw = getLayoutInflater().inflate(R.layout.getfilename,null);
-        final EditText filestr = (EditText) tveiw.findViewById(R.id.editText);
-        Button btnaccept = (Button) tveiw.findViewById(R.id.button2);
-
-
-        btnaccept.setOnClickListener(new View.OnClickListener(){
+        btnRec1.setOnClickListener(new View.OnClickListener(){
             @Override
-            public void onClick(View tveiw){
-                if (!filestr.getText().toString().isEmpty()){
-                    Toast.makeText(MainActivity.this, "Accepted", Toast.LENGTH_SHORT).show();
-                    accepted[0] = filestr.toString();
-                }
+            public void onClick(View v){
+                Log.i("Test","record");
             }
-
         });
 
-        mBuilder.setView(tveiw);
-        AlertDialog dialog = mBuilder.create();
-        dialog.show();
+        ImageButton  imbtnHilt = (ImageButton) findViewById(R.id.imageButton3);
 
-        return filestr.toString();
-    }
-
-
-    class RecordButton extends android.support.v7.widget.AppCompatButton {
-        boolean mStartRecording = true;
-
-        OnClickListener clicker = new OnClickListener() {
-            public void onClick(View v) {
-                onRecord(mStartRecording);
-                if (mStartRecording) {
-                    setText("Stop recording");
-                } else {
-                    setText("Start recording");
-                }
-                mStartRecording = !mStartRecording;
+        imbtnHilt.setOnClickListener(new View.OnClickListener(){
+            @Override
+            public void onClick(View v){
+                imBlade1.setVisibility(VISIBLE);
+                imBlade2.setVisibility(VISIBLE);
             }
-        };
+        });
 
-        public RecordButton(Context ctx) {
-            super(ctx);
-            setText("Start recording");
-            setOnClickListener(clicker);
-        }
+
     }
-
-    private void startRecording() {
-        mRecorder = new MediaRecorder();
-        mRecorder.setAudioSource(MediaRecorder.AudioSource.MIC);
-        mRecorder.setOutputFormat(MediaRecorder.OutputFormat.THREE_GPP);
-        mRecorder.setOutputFile(mFileName);
-        mRecorder.setAudioEncoder(MediaRecorder.AudioEncoder.AMR_NB);
-        Log.e(LOG_TAG, "recording");
-
-        try {
-            mRecorder.prepare();
-        } catch (IOException e) {
-            Log.e(LOG_TAG, "prepare() failed");
-        }
-
-        mRecorder.start();
-    }
-
-    private void stopRecording() {
-        mRecorder.stop();
-        mRecorder.release();
-        mRecorder = null;
-    }
-
 
     @Override
-    public void onCreate(Bundle icicle) {
-        super.onCreate(icicle);
-
-        // Record to the external cache directory for visibility
-
-
-        Log.i(LOG_TAG,"This Far 1");
-
-        ActivityCompat.requestPermissions(this, permissions, REQUEST_RECORD_AUDIO_PERMISSION);
-
-        LinearLayout ll = new LinearLayout(this);
-        mRecordButton = new RecordButton(this);
-        ll.addView(mRecordButton,
-                new LinearLayout.LayoutParams(
-                        ViewGroup.LayoutParams.WRAP_CONTENT,
-                        ViewGroup.LayoutParams.WRAP_CONTENT,
-                        0));
-
-        setContentView(ll);
-
-
+    public boolean onCreateOptionsMenu(Menu menu) {
+        // Inflate the menu; this adds items to the action bar if it is present.
+        getMenuInflater().inflate(R.menu.menu_main, menu);
+        return true;
     }
-
-    public void recordHit() {
-        onRecord(recordable);
-        recordable = !recordable;
-    }
-
 
     @Override
-    public void onStop() {
-        super.onStop();
-        if (mRecorder != null) {
-            mRecorder.release();
-            mRecorder = null;
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
+
+        //noinspection SimplifiableIfStatement
+        if (id == R.id.action_settings) {
+            return true;
         }
+
+        if (super.onOptionsItemSelected(item)) return true;
+        else return false;
     }
+
+
 }
-
-
-
-
