+++ MySaber/app/src/main/res/layout/activity_main.xml
@@ -1,20 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-
-
-
-
-
 <android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@android:color/black"
-    tools:context="hillnerds.me.mysaber.MainActivity"
-    android:exported="true">
-
-
-
+    tools:context="hillnerds.me.mysaber.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
@@ -31,6 +21,5 @@
     </android.support.design.widget.AppBarLayout>
 
     <include layout="@layout/content_main" />
-
 
 </android.support.design.widget.CoordinatorLayout>
