+++ MySaber/app/src/main/res/layout/content_main.xml
@@ -10,41 +10,61 @@
     tools:layout_editor_absoluteY="81dp"
     tools:layout_editor_absoluteX="0dp">
 
+    <ImageView
+        android:id="@+id/imageView"
+        android:layout_width="157dp"
+        android:layout_height="206dp"
+        android:visibility="visible"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:srcCompat="@drawable/light"
+        app:layout_constraintHorizontal_bias="0.502"
+        tools:layout_constraintTop_creator="1"
+        tools:layout_constraintRight_creator="1"
+        tools:layout_constraintLeft_creator="1"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <ImageView
+        android:id="@+id/imageView2"
+        android:layout_width="157dp"
+        android:layout_height="206dp"
+        android:visibility="visible"
+        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintRight_toRightOf="parent"
+        app:srcCompat="@drawable/light"
+        app:layout_constraintHorizontal_bias="0.502"
+        tools:layout_constraintTop_creator="1"
+        tools:layout_constraintRight_creator="1"
+        tools:layout_constraintBottom_creator="1"
+        app:layout_constraintBottom_toBottomOf="parent"
+        tools:layout_constraintLeft_creator="1"
+        app:layout_constraintTop_toTopOf="parent" />
+
     <ImageButton
         android:id="@+id/imageButton3"
-        android:layout_width="118dp"
-        android:layout_height="298dp"
+        android:layout_width="119dp"
+        android:layout_height="399dp"
         android:contentDescription="@string/app_name"
-        android:visibility="visible"
-        app:layout_constraintHorizontal_bias="0.876"
-        app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintRight_toLeftOf="@+id/button"
+        app:layout_constraintLeft_toRightOf="@+id/imageView2"
+        app:layout_constraintRight_toLeftOf="@+id/imageView2"
         app:srcCompat="@drawable/hilt"
-        tools:layout_editor_absoluteY="323dp" />
+        app:layout_constraintHorizontal_bias="0.543"
+        tools:layout_constraintTop_creator="1"
+        android:layout_marginTop="182dp"
+        app:layout_constraintTop_toTopOf="@+id/imageView2" />
 
     <Button
         android:id="@+id/button"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginBottom="179dp"
-        android:layout_marginEnd="16dp"
-        android:layout_marginRight="-525dp"
         android:contentDescription="@string/btnRec1"
-        android:onClick="recordHit()"
         android:text="@string/btnRec1"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintRight_toRightOf="parent"
+        tools:layout_constraintRight_creator="1"
         tools:layout_constraintBottom_creator="1"
-        tools:layout_constraintRight_creator="1" />
-
-    <ImageView
-        android:id="@+id/imageView"
-        android:layout_width="85dp"
-        android:layout_height="374dp"
-        android:visibility="visible"
-        app:layout_constraintLeft_toLeftOf="parent"
+        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_marginEnd="16dp"
         app:layout_constraintRight_toRightOf="parent"
-        app:srcCompat="@drawable/light"
-        tools:layout_editor_absoluteY="-51dp" />
+        android:layout_marginBottom="75dp"
+        android:layout_marginRight="16dp" />
 
 </android.support.constraint.ConstraintLayout>
