--- myapp/gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,0 @@
-#Sat Mar 04 18:32:13 GMT 2017
-distributionBase=GRADLE_USER_HOME
-distributionPath=wrapper/dists
-zipStoreBase=GRADLE_USER_HOME
-zipStorePath=wrapper/dists
-distributionUrl=https\://services.gradle.org/distributions/gradle-3.3-all.zip
